Fix `net.IP` encoding (#4)
Go uses an IPv6-length slice of bytes under the hood by default for
`net.IP` values. When encoding, we now force this to an IPv4 address so
that it only writes the correct number of bytes.

In addition, when writing a slice of bytes, we pad it with zeros if it's
too small.